+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 0.8.0 (unpublished)
 ...................
 * **Breaking change**: Changed CLI command format. Before: ``darq some_project.darq_app.darq``. Now: ``darq -A some_project.darq_app.darq worker``
+* **Breaking change**: Scheduler (cron jobs) now run's seperate from worker (see ``darq scheduler`` command)
+* **Breaking change**: Changed some function signatures (rename arguments)
 
 0.7.2 (2020-06-18)
 ...................